requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
yfinance==0.2.36
openpyxl==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
win10toast==0.9; sys_platform == 'win32'
